added Preview, added UI loading from XML, added XSD schema, added Shadow, added StringView, updated input and UI to use fixed time step, fixed window closing, added drag&drop support, added ObservableList, moved text styles from Theme to a map in Theme, added UI tree logging

+++ Engine/src/ui/drawable/SvgImageDrawable.h
@@ -12,7 +12,7 @@ namespace Ghurund::UI {
         SvgImageDrawable(SvgImage* image) {
             this->image = image;
             image->addReference();
-            preferredSize = { (float)image->Size.width, (float)image->Size.height };
+            preferredSize = { image->Size.width, image->Size.height };
         }
 
         virtual void onDraw(Canvas& canvas) override {
